Prepare for packaging
