--- data/base_data_loader.py
@@ -1,13 +1,0 @@
-
-class BaseDataLoader():
-    def __init__(self):
-        pass
-    
-    def initialize(self):
-        pass
-
-    def load_data():
-        return None
-
-        
-
