+++ perceptual_similarity/data/dataset/base_dataset.py
@@ -1,12 +1,12 @@
-import torch.utils.data as data
+from torch.utils.data import Dataset
 
-class BaseDataset(data.Dataset):
+class BaseDataset(Dataset):
     def __init__(self):
         super(BaseDataset, self).__init__()
-        
+
     def name(self):
         return 'BaseDataset'
-    
+
     def initialize(self):
         pass
