+++ perceptual_similarity/data/dataset/jnd_dataset.py
@@ -1,7 +1,7 @@
 import os.path
 import torchvision.transforms as transforms
-from data.dataset.base_dataset import BaseDataset
-from data.image_folder import make_dataset
+from .base_dataset import BaseDataset
+from ..image_folder import make_dataset
 from PIL import Image
 import numpy as np
 import torch
