+++ perceptual_similarity/data/image_folder.py
