+++ perceptual_similarity/models/base_model.py
@@ -5,9 +5,12 @@ from pdb import set_trace as st
 from IPython import embed
 
 class BaseModel():
+    input_ = None
+    save_dir = None
+
     def __init__(self):
-        pass;
-        
+        pass
+
     def name(self):
         return 'BaseModel'
 
@@ -18,14 +21,11 @@ class BaseModel():
     def forward(self):
         pass
 
-    def get_image_paths(self):
-        pass
-
     def optimize_parameters(self):
         pass
 
     def get_current_visuals(self):
-        return self.input
+        return self.input_
 
     def get_current_errors(self):
         return {}
@@ -39,20 +39,5 @@ class BaseModel():
         save_path = os.path.join(path, save_filename)
         torch.save(network.state_dict(), save_path)
 
-    # helper loading function that can be used by subclasses
-    def load_network(self, network, network_label, epoch_label):
-        save_filename = '%s_net_%s.pth' % (epoch_label, network_label)
-        save_path = os.path.join(self.save_dir, save_filename)
-        print('Loading network from %s'%save_path)
-        network.load_state_dict(torch.load(save_path))
-
-    def update_learning_rate():
+    def update_learning_rate(self):
         pass
-
-    def get_image_paths(self):
-        return self.image_paths
-
-    def save_done(self, flag=False):
-        np.save(os.path.join(self.save_dir, 'done_flag'),flag)
-        np.savetxt(os.path.join(self.save_dir, 'done_flag'),[flag,],fmt='%i')
-
