+++ perceptual_similarity/models/dist_model.py
@@ -18,8 +18,8 @@ from tqdm import tqdm
 
 from IPython import embed
 
-from . import networks_basic as networks
-import models as util
+from .networks_basic import PNetLin, L2, DSSIM, BCERankingLoss, print_network
+from ..util import tensor2im, voc_ap
 
 class DistModel(BaseModel):
     def name(self):
@@ -59,7 +59,7 @@ class DistModel(BaseModel):
         self.model_name = '%s [%s]'%(model,net)
 
         if(self.model == 'net-lin'): # pretrained net + linear layer
-            self.net = networks.PNetLin(pnet_rand=pnet_rand, pnet_tune=pnet_tune, pnet_type=net,
+            self.net = PNetLin(pnet_rand=pnet_rand, pnet_tune=pnet_tune, pnet_type=net,
                 use_dropout=True, spatial=spatial, version=version, lpips=True)
             kw = {}
             if not use_gpu:
@@ -73,12 +73,12 @@ class DistModel(BaseModel):
                 self.net.load_state_dict(torch.load(model_path, **kw), strict=False)
 
         elif(self.model=='net'): # pretrained network
-            self.net = networks.PNetLin(pnet_rand=pnet_rand, pnet_type=net, lpips=False)
+            self.net = PNetLin(pnet_rand=pnet_rand, pnet_type=net, lpips=False)
         elif(self.model in ['L2','l2']):
-            self.net = networks.L2(use_gpu=use_gpu,colorspace=colorspace) # not really a network, only for testing
+            self.net = L2(use_gpu=use_gpu,colorspace=colorspace) # not really a network, only for testing
             self.model_name = 'L2'
         elif(self.model in ['DSSIM','dssim','SSIM','ssim']):
-            self.net = networks.DSSIM(use_gpu=use_gpu,colorspace=colorspace)
+            self.net = DSSIM(use_gpu=use_gpu,colorspace=colorspace)
             self.model_name = 'SSIM'
         else:
             raise ValueError("Model [%s] not recognized." % self.model)
@@ -87,7 +87,7 @@ class DistModel(BaseModel):
 
         if self.is_train: # training mode
             # extra network on top to go from distances (d0,d1) => predicted human judgment (h*)
-            self.rankLoss = networks.BCERankingLoss()
+            self.rankLoss = BCERankingLoss()
             self.parameters += list(self.rankLoss.net.parameters())
             self.lr = lr
             self.old_lr = lr
@@ -103,7 +103,7 @@ class DistModel(BaseModel):
 
         if(printNet):
             print('---------- Networks initialized -------------')
-            networks.print_network(self.net)
+            print_network(self.net)
             print('-----------------------------------------------')
 
     def forward(self, in0, in1, retPerLayer=False):
@@ -180,9 +180,9 @@ class DistModel(BaseModel):
     def get_current_visuals(self):
         zoom_factor = 256/self.var_ref.data.size()[2]
 
-        ref_img = util.tensor2im(self.var_ref.data)
-        p0_img = util.tensor2im(self.var_p0.data)
-        p1_img = util.tensor2im(self.var_p1.data)
+        ref_img = tensor2im(self.var_ref.data)
+        p0_img = tensor2im(self.var_p0.data)
+        p1_img = tensor2im(self.var_p1.data)
 
         ref_img_vis = zoom(ref_img,[zoom_factor, zoom_factor, 1],order=0)
         p0_img_vis = zoom(p0_img,[zoom_factor, zoom_factor, 1],order=0)
@@ -279,6 +279,6 @@ def score_jnd_dataset(data_loader, func, name=''):
 
     precs = TPs/(TPs+FPs)
     recs = TPs/(TPs+FNs)
-    score = util.voc_ap(recs,precs)
+    score = voc_ap(recs,precs)
 
     return(score, dict(ds=ds,sames=sames))
