+++ perceptual_similarity/models/weights/v0.0/alex.pth
