+++ perceptual_similarity/models/weights/v0.0/squeeze.pth
