+++ perceptual_similarity/models/weights/v0.1/alex.pth
