+++ perceptual_similarity/perceptual_loss.py
@@ -3,19 +3,19 @@ from __future__ import absolute_import
 
 import sys
 import scipy
-import scipy.misc
+from imageio import imread
 import numpy as np
 import torch
 from torch.autograd import Variable
-import models
+from . import models
 
 use_gpu = True
 
 ref_path  = './imgs/ex_ref.png'
 pred_path = './imgs/ex_p1.png'
 
-ref_img = scipy.misc.imread(ref_path).transpose(2, 0, 1) / 255.
-pred_img = scipy.misc.imread(pred_path).transpose(2, 0, 1) / 255.
+ref_img = imread(ref_path).transpose(2, 0, 1) / 255.
+pred_img = imread(pred_path).transpose(2, 0, 1) / 255.
 
 # Torchify
 ref = Variable(torch.FloatTensor(ref_img)[None,:,:,:])
@@ -40,15 +40,13 @@ for i in range(1000):
     dist.backward()
     optimizer.step()
     pred.data = torch.clamp(pred.data, 0, 1)
-    
+
     if i % 10 == 0:
         print('iter %d, dist %.3g' % (i, dist.view(-1).data.cpu().numpy()[0]))
         pred_img = pred[0].data.cpu().numpy().transpose(1, 2, 0)
         pred_img = np.clip(pred_img, 0, 1)
-        ax = fig.add_subplot(132)            
+        ax = fig.add_subplot(132)
         ax.imshow(pred_img)
         ax.set_title('iter %d, dist %.3f' % (i, dist.view(-1).data.cpu().numpy()[0]))
         plt.pause(5e-2)
         # plt.imsave('imgs_saved/%04d.jpg'%i,pred_img)
-
-
